move input tests to tests/test_cli
