+++ tests/test_cli/test_input/conftest.py
@@ -1,0 +1,11 @@
+"""Pytest fixtures for tests.test_cli.test_input"""
+from unittest.mock import patch
+
+import pytest
+
+
+@pytest.fixture
+def mock_input_executor():
+    """Prepare a mock input prompt"""
+    with patch("InquirerPy.prompts.input.InputPrompt.execute") as mock_executor:
+        yield mock_executor
