+++ tests/test_cli/test_input/test_select.py
@@ -1,8 +1,12 @@
-import pytest
+"""Tests for redbrick.cli.input.select"""
 from unittest.mock import patch
+
+import pytest
+
 from redbrick.cli.input.select import CLIInputSelect
 
 
+@pytest.mark.unit
 def test_filtrator():
     """Test CLIInputSelect filtrator"""
     cli_input_select = CLIInputSelect(None, "test", ["1", "2", "3"])
@@ -10,6 +14,7 @@ def test_filtrator():
     assert result == "2"
 
 
+@pytest.mark.unit
 def test_validator_options():
     """Test CLIInputSelect validator"""
     cli_input_select = CLIInputSelect(None, "test", ["1", "2", "3"])
@@ -17,6 +22,7 @@ def test_validator_options():
     assert result is True
 
 
+@pytest.mark.unit
 def test_validator_options_false():
     """Test CLIInputSelect validator with input not part of options"""
     cli_input_select = CLIInputSelect(None, "test", ["1", "2", "3"])
@@ -24,6 +30,7 @@ def test_validator_options_false():
     assert result is False
 
 
+@pytest.mark.unit
 def test_validator_dict_options():
     """Test CLIInputSelect validator with input in options"""
     cli_input_select = CLIInputSelect(None, "test", [{"name": "a"}, {"name": "b"}])
@@ -31,6 +38,7 @@ def test_validator_dict_options():
     assert result is True
 
 
+@pytest.mark.unit
 def test_validator_dict_options_false():
     """Test CLIInputSelect validator with input not in options"""
     cli_input_select = CLIInputSelect(None, "test", [{"name": "a"}, {"name": "b"}])
@@ -38,20 +46,26 @@ def test_validator_dict_options_false():
     assert result is False
 
 
+@pytest.mark.unit
 def test_get_from_args():
+    """Test `CLIInputSelect.get` from args"""
     cli_input_select = CLIInputSelect("2", "test", ["1", "2", "3"])
     result = cli_input_select.get()
     assert result == "2"
 
 
+@pytest.mark.unit
 def test_get_from_prompt():
+    """Test `CLIInputSelect.get` from prompt"""
     cli_input_select = CLIInputSelect(None, "test", ["1", "2", "3"])
     with patch("InquirerPy.prompts.fuzzy.FuzzyPrompt.execute", return_value="2"):
         result = cli_input_select.get()
         assert result == "2"
 
 
+@pytest.mark.unit
 def test_get_from_prompt_no_options():
+    """Test `CLIInputSelect.get` with no options"""
     cli_input_select = CLIInputSelect(None, "test", [])
     with pytest.raises(ValueError, match="No test available"):
         cli_input_select.get()
